+++ tasks/serve.js
@@ -28,7 +28,9 @@ exports.default = Task.extend({
         if (projectConfig.project && projectConfig.project.ejected) {
             throw new SilentError('An ejected project cannot use the build command anymore.');
         }
-        rimraf.sync(path.resolve(this.project.root, outputPath));
+        if (serveTaskOptions.deleteOutputPath) {
+            rimraf.sync(path.resolve(this.project.root, outputPath));
+        }
         const serveDefaults = {
             // default deployUrl to '' on serve to prevent the default from .angular-cli.json
             deployUrl: ''
@@ -37,12 +39,19 @@ exports.default = Task.extend({
         let webpackConfig = new webpack_config_1.NgCliWebpackConfig(serveTaskOptions, appConfig).buildConfig();
         const serverAddress = url.format({
             protocol: serveTaskOptions.ssl ? 'https' : 'http',
-            hostname: serveTaskOptions.host,
+            hostname: serveTaskOptions.host === '0.0.0.0' ? 'localhost' : serveTaskOptions.host,
             port: serveTaskOptions.port.toString()
         });
+        if (serveTaskOptions.disableHostCheck) {
+            ui.writeLine(common_tags_1.oneLine `
+          ${chalk.yellow('WARNING')} Running a server with --disable-host-check is a security risk.
+          See https://medium.com/webpack/webpack-dev-server-middleware-security-issues-1489d950874a
+          for more information.
+        `);
+        }
         let clientAddress = serverAddress;
-        if (serveTaskOptions.liveReloadClient) {
-            const clientUrl = url.parse(serveTaskOptions.liveReloadClient);
+        if (serveTaskOptions.publicHost) {
+            const clientUrl = url.parse(serveTaskOptions.publicHost);
             // very basic sanity check
             if (!clientUrl.host) {
                 return Promise.reject(new SilentError(`'live-reload-client' must be a full URL.`));
@@ -67,7 +76,6 @@ exports.default = Task.extend({
                 ui.writeLine('  for information on working with HMR for Webpack.');
                 entryPoints.push('webpack/hot/dev-server');
                 webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
-                webpackConfig.plugins.push(new webpack.NamedModulesPlugin());
                 if (serveTaskOptions.extractCss) {
                     ui.writeLine(common_tags_1.oneLine `
             ${chalk.yellow('NOTICE')} (HMR) does not allow for CSS hot reload when used
@@ -137,14 +145,23 @@ exports.default = Task.extend({
                 poll: serveTaskOptions.poll
             },
             https: serveTaskOptions.ssl,
-            overlay: serveTaskOptions.target === 'development',
-            contentBase: false
+            overlay: {
+                errors: serveTaskOptions.target === 'development',
+                warnings: false
+            },
+            contentBase: false,
+            public: serveTaskOptions.publicHost,
+            disableHostCheck: serveTaskOptions.disableHostCheck
         };
         if (sslKey != null && sslCert != null) {
             webpackDevServerConfiguration.key = sslKey;
             webpackDevServerConfiguration.cert = sslCert;
         }
         webpackDevServerConfiguration.hot = serveTaskOptions.hmr;
+        // set publicPath property to be sent on webpack server config
+        if (serveTaskOptions.deployUrl) {
+            webpackDevServerConfiguration.publicPath = serveTaskOptions.deployUrl;
+        }
         if (serveTaskOptions.target === 'production') {
             ui.writeLine(chalk.red(common_tags_1.stripIndents `
         ****************************************************************************************
@@ -157,7 +174,8 @@ exports.default = Task.extend({
         }
         ui.writeLine(chalk.green(common_tags_1.oneLine `
       **
-      NG Live Development Server is running on ${serverAddress}
+      NG Live Development Server is listening on ${serveTaskOptions.host}:${serveTaskOptions.port},
+      open your browser on ${serverAddress}
       **
     `));
         const server = new WebpackDevServer(webpackCompiler, webpackDevServerConfiguration);
